:art: Reimplemented orientation prop for settings menu
* It had to be reimplemented as the previous version did not work with the layout animations

* 🐛 Fixed bug with playIndicator disappearing after end of scale animation in safari
  * Adding the will-change css property and setting its value to transform fixed it
